+++ debug.log
@@ -1,0 +1,2 @@
+[0830/121428.636:ERROR:icu_util.cc(252)] Couldn't mmap icu data file
+[0830/121428.671:ERROR:icu_util.cc(252)] Couldn't mmap icu data file
